refactor: Config 파일 수정
